updated installer so it won't overwrite vo / adapter configs

+++ install.rb
@@ -2,9 +2,16 @@
 begin
   require 'fileutils'
   Dir.mkdir("./config/rubyamf/")
+  
+  if !File.exist?('./config/rubyamf/vo_config.rb')
+    FileUtils.copy_file("./vendor/plugins/rubyamf/rubyamf_core/rails_installer_files/vo_config.rb", "./config/rubyamf/vo_config.rb", false)
+  end
+  
+  if !File.exist?('./config/rubyamf/adapters_config')
+    FileUtils.copy_file("./vendor/plugins/rubyamf/rubyamf_core/rails_installer_files/adapters_config.rb", "./config/rubyamf/adapters_config.rb", false)
+  end
+  
   FileUtils.copy_file("./vendor/plugins/rubyamf/rubyamf_core/rails_installer_files/rubyamf_controller.rb","./app/controllers/rubyamf_controller.rb",false)
-  FileUtils.copy_file("./vendor/plugins/rubyamf/rubyamf_core/rails_installer_files/vo_config.rb", "./config/rubyamf/vo_config.rb", false)
-  FileUtils.copy_file("./vendor/plugins/rubyamf/rubyamf_core/rails_installer_files/adapters_config.rb", "./config/rubyamf/adapters_config.rb", false)
   FileUtils.copy_file("./vendor/plugins/rubyamf/rubyamf_core/rails_installer_files/crossdomain.xml","./public/crossdomain.xml", false)
   
   mime = true
